requirements.txt
pyautogui
pyaudio
google-cloud-speech